Improved plugin injection
